move several model types
