+++ Mimir/Controllers/AvatarController.cs
@@ -2,8 +2,7 @@ using System.Numerics;
 using Bencodex;
 using Libplanet.Common;
 using Microsoft.AspNetCore.Mvc;
-using Mimir.Models.Agent;
-using Mimir.Models.Avatar;
+using Mimir.Models;
 using Mimir.Repositories;
 using Mimir.Services;
 using Mimir.Util;
