+++ Mimir/Models/AvatarsResponse.cs
@@ -1,4 +1,4 @@
-namespace Mimir.Models.Agent;
+namespace Mimir.Models;
 
 public record AvatarsResponse(
     List<Avatar> Avatars
