+++ Mimir/Repositories/CpRepository.cs
@@ -1,14 +1,11 @@
-using Bencodex.Types;
 using Libplanet.Crypto;
+using Mimir.Models;
 using Nekoyume.Battle;
 using Nekoyume.Model.Item;
 using Nekoyume.Model.Stat;
-using Nekoyume.Model.State;
 using Nekoyume.TableData;
-using Mimir.Models.Agent;
 using Mimir.Services;
 using Mimir.Util;
-using Nekoyume;
 using Nekoyume.Helper;
 using Nekoyume.TableData.Rune;
